parser/tests/tree_construction/basic.dat

#data
<noscript><!--</noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,39): unexpected-end-tag
#script-on
#document
| <html>
|   <head>
|     <noscript>
|       "<!--"
|   <body>
|     "-->"
